+++ docs/src/components/component-example.js
@@ -51,7 +51,7 @@ export default class ComponentExample extends React.Component {
         </div>
         {this.renderCode()}
         <div
-          className={`border border--gray-light px24 py24 ${
+          className={`component-example-${props.title.toLowerCase()} border border--gray-light px24 py24 ${
             this.state.showCode ? 'round-b border-t--0' : 'round-b round-tl '
           }`}
         >