+++ src/components/search/__tests__/search.test.js
@@ -99,22 +99,6 @@ describe('search', () => {
     });
   });
 
-  describe(testCases.withLayout.description, () => {
-    let testCase;
-    let wrapper;
-    let tree;
-
-    beforeEach(() => {
-      testCase = testCases.withLayout;
-      wrapper = renderer.create(testCase.element);
-      tree = wrapper.toJSON();
-    });
-
-    test('renders as expected', () => {
-      expect(tree).toMatchSnapshot();
-    });
-  });
-
   describe(testCases.withConnector.description, () => {
     let testCase;
     let wrapper;