+++ src/helpers/batfish/index.js
@@ -1,0 +1,9 @@
+const { buildNavigation } = require('./navigation');
+const { buildTopics } = require('./topics');
+const { buildSplitPages } = require('./split-pages');
+
+module.exports = {
+  buildNavigation,
+  buildTopics,
+  buildSplitPages
+};