+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # Changelog
 
-## Main
+## 1.0.0
 
+Introducing layouts in `PageLayout`. See the [`PageLayout`](http://mapbox.github.io/dr-ui/guides/page-layout/) guide for more information on building with this component.
+
+- Add `AnalyticsShell` component. [#307](https://github.com/mapbox/dr-ui/pull/307)
+- Add Batfish helpers: `navigation` and `topics`. [#310](https://github.com/mapbox/dr-ui/pull/310)
+- Add Batfish helper: `split-pages`. [#317](https://github.com/mapbox/dr-ui/pull/317)
+- Update `PageLayout` to accept layouts. [#309](https://github.com/mapbox/dr-ui/pull/309)
+  - The added layouts: `page`, `accordion`, `example`, `full`, `exampleIndex`.
+  - 🚨 The following props are deprecated and are no longer configurable: `sidebarTitle`, `sidebarContent`, `sidebarContentStickyTop`, `sidebarContentStickyTopNarrow`, `sidebarStackedOnNarrowScreens`, `sideBarColSize`, `interactiveClass`.
+  - `sidebarTheme` is now defined from the `frontMatter` object.
+  - Component now includes `TopbarSticker/Topbar`, `Search`, and `Feedback` components.
+- 🚨 The `user` props is deprecated in `Feedback`. The component will now fetch this data and you can safely remove the `user` prop from this component. [#324](https://github.com/mapbox/dr-ui/pull/324)
 - Update dependencies. [#326](https://github.com/mapbox/dr-ui/pull/326)
+- Add `cardColSize` prop to `CardContainer` to adjust size of cards.
 
 ## 0.30.0